% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{App}
\usage{
app(
  supplemental_data = rlang::zap(),
  calc_similarity_fn = rlang::zap(),
  n_simulations = 100,
  id_game_vec = rlang::zap()
)
}
\arguments{
\item{supplemental_data}{Optional data joined to required assembled data, using \code{left_join}.}

\item{calc_similarity_fn}{Optional function to calculate similarity scores. The function takes the final dataset (supplemental joined) and must return the same dataset with a new column titled \code{sim_score}.}

\item{n_simulations}{Number of simulations to run per game.}

\item{id_game_vec}{Optional vector of games to limit the simulations to. Defaults to \code{NULL}, which runs all games.}

\item{join_args}{Optional args passed to \code{left_join} joining the supplemental data. This is a required argument if any supplemental data is passed. The relationship must be "one-to-one" and dplyr will throw an error if not.}
}
\value{

}
\description{
Main entry point for the app.
}
