% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastr-derive.R
\name{fastr_derive_fg_metrics}
\alias{fastr_derive_fg_metrics}
\title{Derive Metrics from fastr Data}
\usage{
fastr_derive_fg_metrics(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Derive Metrics from fastr Data
}
