% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fastr.R
\name{fastr_fix_colnames}
\alias{fastr_fix_colnames}
\title{Fix nflfastR Data}
\usage{
fastr_fix_colnames(data, ...)
}
\arguments{
\item{...}{For \code{fastr_fix_columns}, extra columns passed to the renaming function.}
}
\value{
Fixed data.
}
\description{
Fix nflfastR Data
}
