% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_game_team_data}
\alias{get_game_team_data}
\title{Functions For Getting Raw Data}
\usage{
get_game_team_data(seasons)
}
\arguments{
\item{seasons}{int vector of season of which to select for.}
}
\value{

}
\description{
Functions For Getting Raw Data
}
