% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchup-functions.R
\name{prep_matchup}
\alias{prep_matchup}
\title{Functions to Prep and Build the Matchup Profile}
\usage{
prep_matchup(game_data, cur_season, cur_week)
}
\value{

}
\description{
Functions to Prep and Build the Matchup Profile
}
