% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fastr.R
\name{fastr_fix_colnames}
\alias{fastr_fix_colnames}
\title{Reusable Functions Fixing Fastr Data}
\usage{
fastr_fix_colnames(data, ...)
}
\arguments{
\item{...}{For \code{fastr_fix_columns}, extra columns passed to the renaming function.}
}
\value{
Fixed data.
}
\description{
\itemize{
\item \code{fastr_derive_defteam} assigns a defteam to each row. For example, if each row is a game-team combo, we can assign the defteam as the "other" team in the game group.
}
}
